Update connection strings, URLs, and printer settings
Updated the `DefaultConnection` string in `appsettings.Development.json` to point to a new database host (`192.168.100.25`) and database name (`AvyyanKnitfab`), replacing the old host and database name.

Modified the `TallyUrl` in `appsettings.Development.json` under `TallySettings` to use `http://192.168.100.26:9000` instead of `http://192.168.1.23:9001`. The old URL was commented out, and the new URL was uncommented.

Updated the `TallyUrl` in `appsettings.json` under `TallySettings` to `192.168.100.26:9000`.

Changed the `Printer_IP` and `FG_Printer_IP` in `appsettings.json` under `Printers` to `192.168.100.175`, replacing the old IP address `192.168.10.522`.